Make `SegmentIndex` a new type with helpful methods to reduce usage of `PIECES_IN_SEGMENT`

+++ crates/subspace-service/src/dsn/import_blocks/piece_validator.rs
@@ -35,8 +35,11 @@ impl PieceValidator for SegmentCommitmentPieceValidator {
         if source_peer_id != self.dsn_node.id() {
             let segment_index = piece_index.segment_index();
 
-            let maybe_segment_commitment =
-                self.segment_commitment_cache.get(segment_index as usize);
+            // TODO: Needs to statically require 64-bit environment or else u64->usize may cause
+            //  problems in the future
+            let maybe_segment_commitment = self
+                .segment_commitment_cache
+                .get(u64::from(segment_index) as usize);
             let segment_commitment = match maybe_segment_commitment {
                 Some(segment_commitment) => *segment_commitment,
                 None => {
